wave 5: coherence pass; alloc-count bench instrumentation
- helix-modal: split core.rs into core/{key,registry,engine/{helix,vim}}
- helix-plugin: rpc::Rpc -> rpc::Frame (call sites, tests, docs)
- warnings + clippy sweep across workspace (vendor excluded); dead code
  from refactor waves removed; README fork-feature list refreshed
- hx-bench: per-phase allocation counters (calls/bytes for alloc/realloc/
  dealloc); perf round 2 documented in docs/perf-notes.md — all three
  candidates reverted on measured trade-offs

+++ helix-term/src/ui/lsp/hover.rs
@@ -138,7 +138,7 @@ impl Component for Hover {
             0
         });
         let (_, content_height) =
-            crate::ui::text::required_size(&contents_parsed, contents_area.width as u16);
+            crate::ui::text::required_size(&contents_parsed, contents_area.width);
         let scroll_pos = cx.scroll().unwrap_or_default();
         let needs_scrollbar = content_height > contents_area.height;
         // Reserve the rightmost column for the scrollbar when needed
